Shorten IP Address section
The IP address section had and superfluous amount of information which
was not relative to the Server section.

The content was amended and such that it reflected the Server section
more accurately. Adding the notation that servers can and do have IP
Addresses, while linking to DNS information in the extra-reference
directory.